moves to new module "NewApplication"
