+++ .flowconfig
@@ -15,4 +15,4 @@
 module.system.node.resolve_dirname=node_modules
 module.name_mapper='.*\(.s?css\)' -> 'empty/object'
 module.system=haste
-module.name_mapper='\(ActiveDocs\|Applications\|ChangePassword\|Common\|Form\|Settings\|Dashboard\|LoginPage\|Navigation\|Onboarding\|PaymentGateways\|Plans\|Policies\|services\|Stats\|Types\|Users\|utilities\|NewService\)\(.*\)$' -> '<PROJECT_ROOT>/app/javascript/src/\1/\2'
+module.name_mapper='\(ActiveDocs\|Applications\|ChangePassword\|Common\|Form\|Settings\|Dashboard\|LoginPage\|Navigation\|NewApplication\|Onboarding\|PaymentGateways\|Plans\|Policies\|services\|Stats\|Types\|Users\|utilities\|NewService\)\(.*\)$' -> '<PROJECT_ROOT>/app/javascript/src/\1/\2'
