--- app/javascript/src/Applications/index.js
@@ -1,9 +1,0 @@
-// @flow
-
-export * from 'Applications/components/ServicePlanSelect'
-export * from 'Applications/components/ProductFancySelector'
-export * from 'Applications/components/ApplicationPlanSelect'
-export * from 'Applications/components/NameInput'
-export * from 'Applications/components/DescriptionInput'
-export * from 'Applications/components/NewApplicationForm'
-export * from 'Applications/components/NewApplicationFormWrapper'
