+++ app/javascript/src/NewApplication/components/DescriptionInput.jsx
