+++ app/javascript/src/NewApplication/components/NameInput.jsx
