+++ app/javascript/src/NewApplication/components/NewApplicationForm.scss
