+++ app/javascript/src/NewApplication/components/NewApplicationFormWrapper.jsx
@@ -1,6 +1,6 @@
 import React from 'react'
 
-import { NewApplicationForm } from 'Applications'
+import { NewApplicationForm } from 'NewApplication'
 import { createReactWrapper } from 'utilities/createReactWrapper'
 
 const NewApplicationFormWrapper = (props, containerId) => createReactWrapper(<NewApplicationForm {...props} />, containerId)
