+++ app/javascript/src/NewApplication/types.js
